refactor(ui): lift organism business logic into hooks/utils
- BlockGauge: ring geometry, reset countdown, burn-rate/ETA, and live-vs-
  estimate selection move to useBlockGauge; the component is now mostly JSX.
- SettingsView: spending-limit + AI-key draft state and save/clear handlers
  move to useSettingsForm.
- CostCalculation: the cost formula moves to a pure calcCost() in utils.

Behavior unchanged.
