+++ src/components/design-system/organisms/SettingsView/SettingsView.tsx
@@ -1,7 +1,6 @@
-import { useState } from 'react';
 import { ToggleGroup } from '@/components/design-system/atoms/ToggleGroup/ToggleGroup';
-import { parseDollar, fmt } from '@/components/design-system/molecules/LimitsPanel/utils';
 import { PROVIDER_LABELS, PROVIDER_MODELS } from '@/hooks/useAiConfig';
+import { useSettingsForm } from '@/hooks/useSettingsForm';
 import type { Settings } from '@/hooks/useSettings';
 import type { AiProvider } from '@/types';
 import type { SettingsViewProps } from './types';
@@ -18,24 +17,11 @@ export function SettingsView({
   aiConfig,
   onChangeAiConfig,
 }: SettingsViewProps) {
-  const [dailyVal, setDailyVal] = useState(fmt(limits.dailyLimit));
-  const [weeklyVal, setWeeklyVal] = useState(fmt(limits.weeklyLimit));
-  const [monthlyVal, setMonthlyVal] = useState(fmt(limits.monthlyLimit));
-  const [aiKey, setAiKey] = useState(aiConfig.apiKey);
-  const [showKey, setShowKey] = useState(false);
-
-  const providers = Object.keys(PROVIDER_LABELS) as AiProvider[];
-
-  function changeProvider(provider: AiProvider) {
-    // Reset the model to the new provider's first option AND clear the saved key:
-    // keys are provider-specific, so carrying one over would send the wrong
-    // credential to the new provider's API (e.g. an Anthropic key to OpenAI).
-    setAiKey('');
-    onChangeAiConfig({ ...aiConfig, provider, model: PROVIDER_MODELS[provider][0], apiKey: '' });
-  }
-  function saveAiKey() {
-    onChangeAiConfig({ ...aiConfig, apiKey: aiKey.trim() });
-  }
+  const {
+    dailyVal, setDailyVal, weeklyVal, setWeeklyVal, monthlyVal, setMonthlyVal,
+    aiKey, setAiKey, showKey, setShowKey, providers,
+    changeProvider, saveAiKey, clearAiKey, saveLimits, clearLimits,
+  } = useSettingsForm({ limits, onChangeLimits, aiConfig, onChangeAiConfig });
 
   const modeOptions: { value: Settings['modeOverride']; label: string }[] = [
     { value: 'auto', label: 'Auto' },
@@ -43,21 +29,6 @@ export function SettingsView({
     { value: 'api', label: 'API' },
   ];
   const detectedLabel = detectedMode === 'api' ? 'API · pay-as-you-go' : 'Subscription';
-
-  function saveLimits() {
-    onChangeLimits({
-      dailyLimit: parseDollar(dailyVal),
-      weeklyLimit: parseDollar(weeklyVal),
-      monthlyLimit: parseDollar(monthlyVal),
-    });
-  }
-
-  function clearLimits() {
-    onChangeLimits({ dailyLimit: null, weeklyLimit: null, monthlyLimit: null });
-    setDailyVal('');
-    setWeeklyVal('');
-    setMonthlyVal('');
-  }
 
   const inputCls = 'w-full bg-transparent px-2 py-2 text-sm text-zinc-200 outline-none';
   const wrapCls =
@@ -198,10 +169,7 @@ export function SettingsView({
           <div className="flex gap-3">
             {aiConfig.apiKey && (
               <button
-                onClick={() => {
-                  setAiKey('');
-                  onChangeAiConfig({ ...aiConfig, apiKey: '' });
-                }}
+                onClick={clearAiKey}
                 className="rounded-lg px-3 py-1.5 text-xs text-zinc-400 hover:text-zinc-200"
               >
                 clear
